+++ src/snowflake/snowpark/modin/plugin/extensions/general_overrides.py
@@ -33,6 +33,7 @@ import pandas
 import pandas.core.common as common
 from modin.pandas import DataFrame, Series
 from modin.pandas.base import BasePandasDataset
+from modin.pandas.utils import is_scalar
 from pandas import IntervalIndex, NaT, Timedelta, Timestamp
 from pandas._libs import NaTType, lib
 from pandas._libs.tslibs import to_offset
@@ -66,10 +67,6 @@ from pandas.util._validators import validate_inclusive
 # add this line to make doctests runnable
 from snowflake.snowpark import modin
 from snowflake.snowpark.modin import pandas as pd  # noqa: F401
-from snowflake.snowpark.modin.pandas.utils import (
-    is_scalar,
-    raise_if_native_pandas_objects,
-)
 from snowflake.snowpark.modin.plugin._internal.telemetry import (
     snowpark_pandas_telemetry_standalone_function_decorator,
 )
@@ -79,6 +76,9 @@ from snowflake.snowpark.modin.plugin._internal.timestamp_utils import (
 from snowflake.snowpark.modin.plugin._typing import ListLike, ListLikeOfFloats
 from snowflake.snowpark.modin.plugin.compiler.snowflake_query_compiler import (
     SnowflakeQueryCompiler,
+)
+from snowflake.snowpark.modin.plugin.extensions.utils import (
+    raise_if_native_pandas_objects,
 )
 from snowflake.snowpark.modin.plugin.utils.error_message import (
     ErrorMessage,
