+++ src/snowflake/snowpark/modin/plugin/utils/numpy_to_pandas.py
@@ -5,8 +5,8 @@ from typing import Any, Optional, Union
 
 import modin.pandas as pd
 from modin.pandas.base import BasePandasDataset
+from modin.pandas.utils import is_scalar
 
-from snowflake.snowpark.modin.pandas.utils import is_scalar
 from snowflake.snowpark.modin.plugin.utils.warning_message import WarningMessage
