+++ tests/integ/modin/test_utils.py
@@ -11,12 +11,12 @@ from snowflake.snowpark._internal.utils import (
     TempObjectType,
     random_name_for_temp_object,
 )
-from snowflake.snowpark.modin.pandas.utils import (
-    ensure_index,
-    try_convert_index_to_native,
-)
 from snowflake.snowpark.modin.plugin._internal.utils import (
     create_ordered_dataframe_with_readonly_temp_table,
+)
+from snowflake.snowpark.modin.plugin.extensions.utils import (
+    ensure_index,
+    try_convert_index_to_native,
 )
 from tests.integ.modin.sql_counter import SqlCounter, sql_count_checker
 from tests.integ.modin.utils import assert_index_equal
